requirements.txt
django
marshmallow
djangorestframework
django-rest-swagger
